Use `ptrdiff_t` for `PushLimit()`, and document its failure condition.
Previously, `PushLimit()` had an undocumented precondition on `size`, which was required to be within the previous limit.  Violating this precondition would trigger an assert.

It turns out we can remove this precondition, and detect it inside `PushLimit()` itself with a simpler check (`delta < 0`).  We now document this failure condition explicitly, and the callers of `PushLimit()` (including the tests) are updated to check for this failure condition.

In a future CL, we will remove the existing eager check on `size`, which is no longer necessary.

PiperOrigin-RevId: 837932824